+++ eqsig/duhamels.py
@@ -1,4 +1,3 @@
-from eqsig.sdof import *
 from eqsig.exceptions import deprecation
 
 deprecation('eqsig.duhamels has deprecated, used eqsig.sdof')